ref: improve providergen structure (#898)
* ref: improve providergen structure

* chore: rename int to int64

* ref: model type

+++ internal/providergen/data_sources/cron_monitor.ts
@@ -76,7 +76,7 @@ export default {
     },
     {
       name: "checkin_margin_minutes",
-      type: "int",
+      type: "int64",
       description:
         "Grace period. The number of minutes before a check-in is considered missed.",
       computedOptionalRequired: "computed",
@@ -84,7 +84,7 @@ export default {
     },
     {
       name: "failure_issue_threshold",
-      type: "int",
+      type: "int64",
       description:
         "Failure tolerance. Create a new issue when this many consecutive missed or error check-ins are processed.",
       computedOptionalRequired: "computed",
@@ -92,7 +92,7 @@ export default {
     },
     {
       name: "max_runtime_minutes",
-      type: "int",
+      type: "int64",
       description:
         "Maximum runtime. The number of minutes before an in-progress check-in is marked timed out.",
       computedOptionalRequired: "computed",
@@ -100,7 +100,7 @@ export default {
     },
     {
       name: "recovery_threshold",
-      type: "int",
+      type: "int64",
       description:
         "Recovery Tolerance. Resolve the issue when this many consecutive healthy check-ins are processed. Either `crontab` or `interval_value` and `interval_unit` must be provided.",
       computedOptionalRequired: "computed",
@@ -122,7 +122,7 @@ export default {
         },
         {
           name: "interval_value",
-          type: "int",
+          type: "int64",
           description: "Interval value for the cron monitor.",
           computedOptionalRequired: "computed",
           skipFill: true,
