+++ internal/providergen/resources/metric_monitor.ts
@@ -143,7 +143,7 @@ export default {
     },
     {
       name: "time_window_seconds",
-      type: "int",
+      type: "int64",
       description: "The time window in seconds to use for the aggregate query.",
       computedOptionalRequired: "computed_optional",
     },
@@ -170,7 +170,7 @@ export default {
         },
         {
           name: "comparison_delta",
-          type: "int",
+          type: "int64",
           description:
             "The comparison delta in seconds to use for the aggregate query. Only required for `percent` type.",
           computedOptionalRequired: "optional",
@@ -244,7 +244,7 @@ export default {
             },
             {
               name: "condition_result",
-              type: "int",
+              type: "int64",
               description:
                 "When the condition is met, the result will be set to this value.",
               computedOptionalRequired: "required",
